+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-05-08T07:26:47.595Z
+Generated: 2026-05-09T07:53:02.123Z
 
 - 2026-07-30 • Microsoft Defender for Cloud • Deprecation • Deprecation notice: Legacy grouped recommendations
 - 2026-07-01 • Microsoft Sentinel • Action required • [Updated] Call to action: update automation by July 1, 2026 - Account Name is now consistently the UPN prefix for analytics rule alerts
@@ -15,10 +15,10 @@ Generated: 2026-05-08T07:26:47.595Z
 - 2026-05-05 • Microsoft Defender for Cloud • Ikke oppgitt • Updated Helm installation for Defender for Containers sensor
 - 2026-05-05 • Microsoft Defender for Cloud • Ikke oppgitt • Daily score calculation enhancement for risk-based Cloud secure score
 - 2026-05-03 • Microsoft Defender for Cloud • GA • General availability of of Defender for Cloud integration into the Defender portal
-- 2026-05-01 • Microsoft Defender for Endpoint • Preview • Selective Response Actions
+- 2026-05-01 • Microsoft Defender for Endpoint • Preview • Configure offline security intelligence update settings for Linux from the Defender and Intune portals
 - 2026-05-01 • Microsoft Defender for Endpoint • Preview • Selective Response Actions enables organizations to tailor high-impact security operations on devices during onboarding. It provides precise control over how response actions are applied on Tier-0 systems and other high-value assets, helping maintain operational stability while delivering strong protection.
+- 2026-05-01 • Microsoft Defender for Endpoint • Ikke oppgitt • You can now configure offline security intelligence update settings for Linux directly from the Defender and Intune portals.
 - 2026-05-01 • Microsoft Defender for Identity • Ikke oppgitt • New Defender for Identity security alerts
 - 2026-05-01 • Microsoft Defender XDR • Ikke oppgitt • In advanced hunting, the
 - 2026-05-01 • Microsoft Entra • Deprecation • Plan for change – Agent Registry consolidation into Microsoft Agent 365
 - 2026-05-01 • Microsoft Purview • GA • General availability (GA)
-- 2026-05-01 • Microsoft Sentinel • Ikke oppgitt • We introduced a new entry point and created a consolidated view for the UEBA Settings and the Behaviors Settings. You can now access the UEBA settings from the new UEBA tab in the Microsoft Sentinel settings page. For more information see